edit

+++ src/component/ListStuff.js
@@ -10,14 +10,21 @@ const Stuff = (props) => (
     <td>{props.stuff.duration.substring(0, 10)}</td>
     <td>{props.stuff.date.substring(0, 10)}</td>
     <td>
-      <Link to={"/edit/" + props.stuff._id}>edit</Link> |{" "}
+      <Link className="edit" to={"/edit/" + props.stuff._id}>
+        edit
+      </Link>{" "}
+      |{" "}
       <a
         href="#"
+        className="delete"
         onClick={() => {
-          props.deleteStuff(props.stuff._id);
+          const result = window.confirm("Hapus barang dari daftar?");
+          if (result) {
+            props.deleteStuff(props.stuff._id);
+          }
         }}
       >
-        delete
+        hapus
       </a>
     </td>
   </tr>
@@ -36,7 +43,7 @@ export default function ListStuff(props) {
       .catch((error) => {
         console.log(error);
       });
-  }, []);
+  });
 
   const deleteStuff = (id) => {
     axios.delete("http://localhost:5000/stuff/" + id).then((response) => {
@@ -61,26 +68,18 @@ export default function ListStuff(props) {
   return (
     <ListToko>
       <div className="list-content">
-        <h3>Toko Sebelah</h3>
-        <p>Daftar Barang di toko Kami</p>
         <table className="table">
           <thead className="thead-light">
             <tr>
               <th>Nama Barang</th>
               <th>Harga</th>
               <th>Kadaluarsa</th>
-              <th>Date</th>
-              <th>Actions</th>
+              <th>Tanggal Masuk</th>
+              <th>Aksi</th>
             </tr>
           </thead>
           <tbody>{stuffList()}</tbody>
         </table>
-      </div>
-
-      <div className="add-stuf">
-        <Link to="/create">
-          <a className="button-add">+ Tambah Barang</a>
-        </Link>
       </div>
     </ListToko>
   );
@@ -88,7 +87,8 @@ export default function ListStuff(props) {
 
 const ListToko = styled.div`
   width: 100%;
-  height: 100vh;
+  min-height: 100vh;
+  padding: 8rem 0 3rem 0;
   display: flex;
   align-items: center;
   flex-direction: column;
@@ -100,16 +100,15 @@ const ListToko = styled.div`
     display: flex;
     align-items: center;
     flex-direction: column;
-    padding: 10vh 0;
   }
 
   .list-content > h3 {
-    font-size: 7vh;
+    font-size: 2rem;
     margin: 0;
   }
 
   .list-content > p {
-    font-size: 3vh;
+    font-size: 1rem;
   }
 
   .list-content > table {
@@ -127,28 +126,50 @@ const ListToko = styled.div`
     justify-content: center;
   }
 
-  tbody {width:100%;}
+  tbody {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+  }
   thead > tr {
     width: 100%;
-    margin-top: 3vh;
+
     border: 1px solid #213936;
-    padding: 2vh 0;
+    padding: 1.5rem 0;
     display: flex;
     flex-direction: row;
     align-items: center;
     justify-content: space-evenly;
   }
 
-  th,td {width:10vw;
- text-align:center;}
-  tr {
+  th,
+  td {
+    width: 15%;
+    text-align: center;
+
+    .edit {
+      &:hover {
+        color: green;
+      }
+    }
+
+    .delete {
+      &:hover {
+        color: red;
+      }
+    }
+  }
+  tbody > tr {
     width: 100%;
-    margin-top: 3vh;
+    height: 5rem;
 
     display: flex;
     flex-direction: row;
     align-items: center;
     justify-content: space-evenly;
+    border-bottom: 1px solid var(--black);
   }
 
   .add-stuf {
@@ -157,18 +178,19 @@ const ListToko = styled.div`
     justify-content: flex-end;
     margin-top: 5vh;
   }
-a {   text-decoration:none;}
+  a {
+    text-decoration: none;
+  }
 
-  .button-add{
-    
-   text-decoration:none;
-   color: black;
-   border: 1px solid black;
-   padding: 2vh 1vh;
+  .button-add {
+    text-decoration: none;
+    color: black;
+    border: 1px solid black;
+    padding: 2vh 1vh;
 
-   &:hover {
+    &:hover {
       background-color: black;
-      color:white;
+      color: white;
     }
   }
 `;
